requirements.txt
Flask==2.3.2
Pillow==10.1.0
numpy==1.27.0
opencv-python-headless==4.8.1.78
gunicorn==20.1.0
